third_party/WebKit/LayoutTests/accessibility/inline-text-changes-expected.txt

This paragraph contains two lines of text. One more sentence.

Tests that accessible inline text boxes update when a static text node changes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS axStaticText.childrenCount is 2
PASS axInlineBefore0.stringValue is 'AXValue: This paragraph contains '
PASS axInlineBefore1.stringValue is 'AXValue: two lines of text.'
PASS axStaticText.childrenCount is 3
PASS axInlineAfter0.stringValue is 'AXValue: This paragraph contains '
PASS axInlineAfter1.stringValue is 'AXValue: two lines of text. One '
PASS axInlineAfter2.stringValue is 'AXValue: more sentence.'
PASS axInlineBefore1.stringValue is 'AXValue: '
PASS successfullyParsed is true

TEST COMPLETE
